DOS driver fixes, test app updates, CMake tweaks
Fix pointer cast in dosdma.c to silence/avoid incorrect pointer conversion on some compilers by casting via size_t. Adjust dosutil.h preprocessor logic to only include <pc.h> for __DJGPP__. Remove built binaries from miktest/BIN. Update miktest CMakeLists: add MIKMOD_STATIC define when linking static, and enforce C99 without compiler extensions. Revise MIKTEST.c: generalize displayed program name, add conditional driver selection (DRV_SB vs nos), change default mix settings (44100Hz for nos), and reorganize MikMod_Init path to match driver selection.

+++ libmikmod/drivers/dos/dosdma.c
@@ -140,7 +140,7 @@ dma_buffer *dma_allocate(unsigned int channel, unsigned int size)
 #ifdef __DJGPP__
 	buffer->linear = (unsigned char *)(__djgpp_conventional_base + bound * 16);
 #else
-	buffer->linear = (unsigned char *)(bound * 16);
+	buffer->linear = (unsigned char *)(size_t)(bound * 16);
 #endif
 	buffer->physical = bound * 16;
 	buffer->size = parsize * 16;
